# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Find Your Car",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/fyc/settings": {
            "get": {
                "description": "Get settings by CarPark ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get settings by CarPark ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CarPark ID",
                        "name": "carpark_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Settings"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing settings by CarPark ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Update settings by CarPark ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CarPark ID",
                        "name": "carpark_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Updated settings data",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Settings"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new settings for a CarPark",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Add new settings",
                "parameters": [
                    {
                        "description": "Settings data",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Settings"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pkg.Settings"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pkg.Settings": {
            "type": "object",
            "required": [
                "app_logo",
                "carpark_id",
                "carpark_name",
                "couting_clean_cron",
                "default_lang",
                "fyc_clean_cron",
                "timeout_screenKiosk"
            ],
            "properties": {
                "app_logo": {
                    "type": "string"
                },
                "carpark_id": {
                    "type": "integer"
                },
                "carpark_name": {
                    "type": "string"
                },
                "couting_clean_cron": {
                    "type": "string"
                },
                "default_lang": {
                    "type": "string",
                    "default": "EN"
                },
                "fyc_clean_cron": {
                    "type": "string"
                },
                "timeout_screenKiosk": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}